# api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/invoices": {
      "get": {
        "operationId": "InvoicesController_list",
        "summary": "List invoices",
        "description": "Paginated list of invoices in the organisation.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Paginated invoices",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": [
                      {
                        "id": "665f1b2c9a4e1c0012ab34cd",
                        "orgId": "665f1a009a4e1c0012ab1100",
                        "type": "export",
                        "clientId": "665f1b2c9a4e1c0012ab34cd",
                        "number": "INV-2026-0007",
                        "financialYear": "2026-27",
                        "invoiceDate": "2026-06-12",
                        "currency": "USD",
                        "fxRate": "83.4500",
                        "subtotal": "5000.00",
                        "inrEquivalent": "417250.00",
                        "taxType": "LUT_ZERO",
                        "grandTotal": "5000.00",
                        "status": "draft",
                        "femaDueDate": "2026-12-09",
                        "pdfUrl": null,
                        "items": [
                          {
                            "description": "Software development services",
                            "sacCode": "998314",
                            "quantity": "1",
                            "unitAmount": "5000.00",
                            "lineTotal": "5000.00",
                            "gstRate": 0
                          }
                        ]
                      }
                    ],
                    "meta": {
                      "page": 1,
                      "limit": 20,
                      "total": 1,
                      "totalPages": 1
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "InvoicesController_create",
        "summary": "Create an invoice",
        "description": "Creates an export or domestic invoice. Number, FX rate, tax breakup and FEMA due date are computed server-side.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invoice created",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "id": "665f1b2c9a4e1c0012ab34cd",
                      "orgId": "665f1a009a4e1c0012ab1100",
                      "type": "export",
                      "clientId": "665f1b2c9a4e1c0012ab34cd",
                      "number": "INV-2026-0007",
                      "financialYear": "2026-27",
                      "invoiceDate": "2026-06-12",
                      "currency": "USD",
                      "fxRate": "83.4500",
                      "subtotal": "5000.00",
                      "inrEquivalent": "417250.00",
                      "taxType": "LUT_ZERO",
                      "grandTotal": "5000.00",
                      "status": "draft",
                      "femaDueDate": "2026-12-09",
                      "pdfUrl": null,
                      "items": [
                        {
                          "description": "Software development services",
                          "sacCode": "998314",
                          "quantity": "1",
                          "unitAmount": "5000.00",
                          "lineTotal": "5000.00",
                          "gstRate": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/invoices/fema/aging": {
      "get": {
        "operationId": "InvoicesController_femaAging",
        "summary": "FEMA aging buckets",
        "description": "Export invoices grouped by days-to-realisation bucket against the FEMA deadline.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Aging buckets",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "buckets": {
                        "0-30": 2,
                        "31-60": 1,
                        "overdue": 0
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/invoices/{id}": {
      "get": {
        "operationId": "InvoicesController_findOne",
        "summary": "Get an invoice by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Invoice id",
            "schema": {
              "example": "665f1b2c9a4e1c0012ab34cd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invoice",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "id": "665f1b2c9a4e1c0012ab34cd",
                      "orgId": "665f1a009a4e1c0012ab1100",
                      "type": "export",
                      "clientId": "665f1b2c9a4e1c0012ab34cd",
                      "number": "INV-2026-0007",
                      "financialYear": "2026-27",
                      "invoiceDate": "2026-06-12",
                      "currency": "USD",
                      "fxRate": "83.4500",
                      "subtotal": "5000.00",
                      "inrEquivalent": "417250.00",
                      "taxType": "LUT_ZERO",
                      "grandTotal": "5000.00",
                      "status": "draft",
                      "femaDueDate": "2026-12-09",
                      "pdfUrl": null,
                      "items": [
                        {
                          "description": "Software development services",
                          "sacCode": "998314",
                          "quantity": "1",
                          "unitAmount": "5000.00",
                          "lineTotal": "5000.00",
                          "gstRate": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invoice not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "patch": {
        "operationId": "InvoicesController_update",
        "summary": "Update invoice status",
        "description": "Only the status may change; financial fields are immutable.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Invoice id",
            "schema": {
              "example": "665f1b2c9a4e1c0012ab34cd",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invoice updated",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "id": "665f1b2c9a4e1c0012ab34cd",
                      "orgId": "665f1a009a4e1c0012ab1100",
                      "type": "export",
                      "clientId": "665f1b2c9a4e1c0012ab34cd",
                      "number": "INV-2026-0007",
                      "financialYear": "2026-27",
                      "invoiceDate": "2026-06-12",
                      "currency": "USD",
                      "fxRate": "83.4500",
                      "subtotal": "5000.00",
                      "inrEquivalent": "417250.00",
                      "taxType": "LUT_ZERO",
                      "grandTotal": "5000.00",
                      "status": "sent",
                      "femaDueDate": "2026-12-09",
                      "pdfUrl": null,
                      "items": [
                        {
                          "description": "Software development services",
                          "sacCode": "998314",
                          "quantity": "1",
                          "unitAmount": "5000.00",
                          "lineTotal": "5000.00",
                          "gstRate": 0
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Invoice not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/invoices/{id}/pdf": {
      "get": {
        "operationId": "InvoicesController_pdf",
        "summary": "Get invoice PDF url",
        "description": "Returns the generated PDF url, or 404 while generation is still queued.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Invoice id",
            "schema": {
              "example": "665f1b2c9a4e1c0012ab34cd",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF location",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "data": {
                      "url": "https://files.crossbill.app/invoices/INV-2026-0007.pdf",
                      "status": "sent"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but role not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "PDF not ready yet; generation is queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Invoices"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Crossbill API",
    "description": "Crossbill backend API — GST / FEMA compliant invoicing & remittance for Indian exporters.\n\n**Conventions**\n- Base path: `/api/v1`\n- Success envelope: `{ \"success\": true, \"data\": ... }` (list endpoints add a `meta` object).\n- Error envelope: `{ \"success\": false, \"error\": { \"code\", \"message\", \"correlationId\" }, \"timestamp\", \"path\" }`.\n- Auth: send `Authorization: Bearer <accessToken>` (obtained from `/auth/login` or `/auth/register`).\n- The refresh token is delivered as the `crossbill_rt` httpOnly cookie and is read automatically by `/auth/refresh`.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Base URL (includes /api/v1 prefix)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the accessToken returned by /auth/login or /auth/register."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "crossbill_rt",
        "description": "httpOnly refresh-token cookie set by the auth endpoints; used by /auth/refresh."
      }
    },
    "schemas": {
...
      "ApiErrorBody": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Validation failed",
            "description": "Human-readable error message"
          },
          "correlationId": {
            "type": "string",
            "example": "7f3c1e9a-2b4d-4c8e-9a1b-1f2e3d4c5b6a",
            "description": "Request correlation id"
          }
        },
        "required": [
          "code",
          "message",
          "correlationId"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-06-12T10:15:30.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/clients"
          }
        },
        "required": [
          "success",
          "error",
          "timestamp",
          "path"
        ]
      },
...
      "InvoiceItemDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Line item description",
            "example": "Software development services - June 2026"
          },
          "sacCode": {
            "type": "string",
            "description": "SAC (Services Accounting Code)",
            "enum": [
              "998313",
              "998314",
              "998315",
              "998316",
              "998319",
              "998361",
              "998391",
              "998399"
            ],
            "example": "998314"
          },
          "quantity": {
            "type": "number",
            "description": "Quantity",
            "example": 1,
            "minimum": 0
          },
          "unitAmount": {
            "type": "number",
            "description": "Unit amount in the invoice currency",
            "example": 5000,
            "minimum": 0
          },
          "gstRate": {
            "type": "number",
            "description": "GST rate % (domestic only). Ignored for export invoices.",
            "enum": [
              0,
              5,
              12,
              18,
              28
            ],
            "example": 18
          }
        },
        "required": [
          "description",
          "quantity",
          "unitAmount"
        ]
      },
      "CreateInvoiceDto": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "description": "Id of an existing client to bill",
            "example": "665f1b2c9a4e1c0012ab34cd"
          },
          "invoiceDate": {
            "type": "string",
            "description": "Invoice date (ISO 8601). Defaults to today when omitted.",
            "format": "date",
            "example": "2026-06-12"
          },
          "currency": {
            "type": "string",
            "description": "Invoice currency. Required for export; domestic invoices are forced to INR by the server.",
            "enum": [
              "USD",
              "EUR",
              "GBP",
              "AUD",
              "CAD",
              "SGD",
              "AED",
              "INR"
            ],
            "example": "USD"
          },
          "items": {
            "description": "At least one line item",
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItemDto"
            }
          }
        },
        "required": [
          "clientId",
          "items"
        ]
      },
      "UpdateInvoiceDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "New invoice status. Financial fields cannot be changed once captured.",
            "enum": [
              "draft",
              "sent",
              "paid",
              "overdue"
            ],
            "example": "sent"
          }
        }
      },
...
    }
  }
}